+++ docs/main/changelog.rst
@@ -11,6 +11,7 @@ Development version
 -------------------
 - Add ``ThreadLocalSingleton`` and ``DelegatedThreadLocalSingleton`` providers.
 - Add documentation section about singleton providers and multi-threading.
+- Update API docs of creational providers.
 
 2.0.0
 ------